pyproject.toml
[tool.poetry]
name = "temperature-alert-agent"
version = "0.1.0"
description = ""
authors = ["A-J-Sarmah <ajsarmah@tutanota.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<3.12"
uagents = "^0.6.2"
poetry-dotenv-plugin = "^0.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
